ENH: Refactor code linter and formatter dependencies
Refactor code linter and formatter dependencies:
- Remove `black`, `flake8` and `isort` dependencies from
  `pyproject.toml`: the role has been taken over by `ruff` in commit
  81b3690.
- Transfer the `black` and `isort` rules left behind to `ruff` in
  `pyproject.toml`
- Remove `black`, `flake8` and `isort` from the pre-commit config file
  and use `ruff` instead.

Set explicitly `eddymotion` as the known first party so that the
`import` statements corresponding to `eddymotion` are separated by a
blank line from third-party package `import` statements.

Add `--diff` to the `ruff check` command in the contribution checks
GitHub Actions workflow so that the differences that need to be applied
are actually shown on GitHub.