BUG: Fix tests
Fix GP model testing: `GaussianProcessModel` is now hosted in the
`model.dipy` module.

Fix bvals array dimensionality and make bvecs be unit vectors in angle
computation test.

Fix the pairwise angle computation doctests.

Test the case where the second argument to the pairwaise angle
computation function is `None`. Simplify the test parameterization as
the former corresponds to providing the same value for both gtab
arguments. Adapt the test so that we avoid making computations if the
second argument is `None`.