MAINT: Adopt `tox` for testing environment configuration
Adopt `tox` for testing environment configuration as per decision at the
Oct 1 2024 Technical Monitoring meeting.

Add the necessary `tox.ini` file and edit the GHA `test.yml` workflow
file to use `tox`, to use `ants`, to use `DataLad` to download the
testing data, and to cache the data across the build matrix.

Pin `nitransforms` to <24 following commit d26541b. The decision on
pinning to <24 was made following consistent failures on
`test_proximity_estimator_trivial_model`, e.g.
https://github.com/nipreps/eddymotion/actions/runs/11123784605/job/30907972182#step:8:196
and
https://github.com/nipreps/eddymotion/actions/runs/11203922080/job/31141711676?pr=234#step:12:342

The version was pinned for the conda `env.yml` config file, but not
`pyproject.toml`.