DOC: Update NOTICE file text
The `NOTICE` file text is not a repetition of the license banner, but
clarifications about the licensing of the code (e.g., if in the past had
other licenses, or if it derives from pre-exiting code).

Thus, remove the license information and edit the text to reflect the
copyright of the code.

Co-authored-by: Oscar Esteban <code@oscaresteban.es>