fix: address normalization of training data
@jhlegarreta @yasseraleman -- I may have found the potential culprit
that makes fitting so weird.

By looking at the covariance vs. angle plots (at the beginning of the
notebook), I ended up realizing that they massaged the data before
calculating the covariance.
They certainly removed the mean (I don't know if gradient-wise mean or
the grand mean), and possibly scaled to one-standard-deviation.

Then I went back to the paper and found the little comment I included
in the docstring of ``EddymotionGPR.fit()``.

This doesn't completely resolve all problems, but I'm positive it's one
little step we needed to take.