enh: finalize PR
It seems I was wrong about the need for standardizing, Scikit-learn's
implementation is sufficient and follows what the paper specifies.

There are a few additional details that have surfaced as relevant:

* $\sigma^2$ should not be very large (and ~100) was VERY large.
  With small values, it seems to pacify optimizers with more plausible
  outcomes.
* $a$ may take values above $\frac{\pi}{2}$, even though $\theta$ will
  always be smaller than that.
  Indeed, it seems for our real dataset the optimal $a$ is indeed around
  $\frac{\pi}{2}$.
* I haven't cracked the manipulations of the covariance plot yet, but
  it's clear to me now that they are indeed considering $a$ and $\lambda$.

The most relevant inclusion in this commit is the repetition of the
covariance plot with the covariance of predicted data, which is very
encouraging.

cc @jhlegarreta.