DOC: Fix cut passages from `README` in documentation index file
Fix cut passages from `README` in documentation index file: add comments
around the flowchart image in the `README` file and change the
`start-line` and `end-line` directive pair for the `end-before` and
`start-after` pair to skip the flowchart image only.

The `start-line` and `end-line` directives were making such that
relevant passages of the `README` file providing context for the
tool were not being included. Thus the text shown in the documentation
index file had some missing parts, and was not making sense.